Achieved symbol automation perfection!
Being mostly done with the 3D parts (for now, until I think of more, just had to do a few thumbnails and maybe finish the negative spherics, can't remember) getting up to symbols and holy cow, there's a lot of variants of one thing I can make.
Automate using OpenSCAD to extrude the SVGs? Check.
Put the STL and copy the SVG (to use as the thumbnail) into the models folder? Check.
Programmatically generate a dictionary for all the symbols I can diff in as required? Heck check!