SVG madness!
To stop cows loading 100mm to the right of the build plate for whatever reason, updated the SVG > STL script to centre and resize the SVG so they're all consistent (requires development snapshot of OpenSCAD).
Also: Been on the warpath hunting down leftover bits of code which no longer have a reason to exist.
*cries* And I thought of another style to do emoji symbols in...